# src/main/resources/api.yaml
openapi: "3.0.2"
info:
  title: API Title
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local testing server
paths:
  /users:
    get:
      description: Returns a list of all users.
      operationId: getUsers
      responses:
        '200':
          description: The list of all users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      description: Creates a new user.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Creates a new user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: string
    get:
      description: Returns a single user.
      operationId: getUser
      responses:
        '200':
          description: The user with the given id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: The user was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: Deletes a single user.
      operationId: deleteUser
      responses:
        '204':
          description: The user has been deleted.
        '404':
          description: The user was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          example: Max Mustermann
          minLength: 1
          maxLength: 255
        birthday:
          type: string
          example: 1993-09-14
          format: date
      required:
        - name
        - birthday
    User:
      type: object
      properties:
        id:
          type: string
          example: 123
          minLength: 1
          maxLength: 36
          pattern: '^[0-9a-z]+$'
        name:
          type: string
          example: Max Mustermann
          minLength: 1
          maxLength: 255
        birthday:
          type: string
          example: 1993-09-14
          format: date
      required:
        - id
        - name
        - birthday
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    Error:
      type: object
      properties:
        code:
          type: string
          pattern: '^[a-z]+(_[a-z]+)*$'
        description:
          type: string
          minLength: 1
          maxLength: 255